:boom: CardEvent -> RailEvent
